+++ src/views/wallet/Operations.vue
@@ -182,15 +182,11 @@ const appOption = useAppOptionStore();
 
 import { useStoreWallet } from "@/stores/wallet.ts";
 const storeWallet = useStoreWallet();
-import typeahead from '@/components/plugins/Typeahead.vue';
-//import vueSelect from '@/components/plugins/vueSelect.vue';
 
 export default {
   name: "OperationsComponent",
   components: {
     QrcodeVue,
-    typeahead,
-    //vueSelect,
   },
   props: {
     address: String,