+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v1.1.1
 - add status transaction Success or Err
 - mobile view transactions list
 - add component list transactions
+- add explorer onclick txId
 
 v1.1.0
 - add cross chain support XBTS DEX (BitShares)